feat : add filterBar to RecipesScreen and make functions FilterBar, FilterChipItem and TopNavigationBar more general
