+++ app/src/main/java/com/android/shelfLife/ui/recipes/IndividualRecipe.kt
@@ -163,7 +163,9 @@ fun IndividualRecipeScreen(
                   },
                   houseHold = it,
                   householdViewModel = householdViewModel,
-                  onHamburgerClick = { scope.launch { drawerState.open() } })
+                  onHamburgerClick = { scope.launch { drawerState.open() } },
+                  filters = emptyList()
+              )
             }
           },
           bottomBar = {
